LayoutTests/imported/w3c/web-platform-tests/encoding/legacy-mb-japanese/shift_jis/sjis-decode-errors-expected.txt


FAIL lead not 0x00 and no more bytes: FA assert_equals: expected "\ufffd" but got "\x1a"
FAIL lead not 0x00 and no more bytes: FA 6E FA assert_equals: expected "佖\ufffd" but got "佖\x1a"
FAIL lead not 0x00 and no more bytes: FA FA FA assert_equals: expected "洄\ufffd" but got "洄\x1a"
FAIL lead byte outside 0x81-0x9F,0xA1-0xDF,0xE0,0xFC: FF assert_equals: expected "\ufffd" but got "\x1a"
FAIL lead byte outside 0x81-0x9F,0xA1-0xDF,0xE0,0xFC: A0 assert_equals: expected "\ufffd" but got "\x1a"
FAIL lead byte outside 0x81-0x9F,0xA1-0xDF,0xE0,0xFC: E1 assert_equals: expected "\ufffd" but got "\x1a"
FAIL lead byte outside 0x81-0x9F,0xA1-0xDF,0xE0,0xFC: FA FA FF assert_equals: expected "洄\ufffd" but got "洄\x1a"
FAIL trail byte outside 0x41-0xFE: FA FD assert_equals: expected "\ufffd" but got "\x1a\x1a"
FAIL trail byte outside 0x41-0xFE: FA FE assert_equals: expected "\ufffd" but got "\x1a\x1a"
PASS pointer is null: 81 B5 
